+++ index.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8">
     <link rel="icon" href="/favicon.ico">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- <link rel="stylesheet" href="./src/assets/output.css"> -->
     <title>Vite App</title>
   </head>
   <body>
