+++ package.json
@@ -25,8 +25,6 @@
     "js-cookie": "^3.0.5",
     "pinia": "^2.2.6",
     "postcss": "^8.4.49",
-    "sass": "^1.82.0",
-    "sass-loader": "^16.0.4",
     "tailwindcss": "^3.4.16",
     "vee-validate": "^4.14.7",
     "vue": "^3.5.13",
