+++ src/main.ts
@@ -1,5 +1,7 @@
 import './assets/main.css'
 import 'vuetify/styles'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
@@ -8,9 +10,6 @@ import { VueQueryPlugin, QueryClient } from '@tanstack/vue-query'
 
 import router from './router'
 import App from './App.vue'
-
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
 
 const app = createApp(App)
 const queryClient = new QueryClient()
