+++ src/router/index.ts
@@ -1,6 +1,8 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import Login_page from '@/views/Login_page.vue'
-import Forgot_password from '@/views/Forgotpassword_page.vue'
+
+import LoginPage from '@/views/LoginPage.vue'
+import ForgotPasswordPage from '@/views/ForgotPasswordPage.vue'
+
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
@@ -8,12 +10,12 @@ const router = createRouter({
     {
       path: '/login',
       name: 'login',
-      component: Login_page,
+      component: LoginPage,
     },
     {
       path: '/forgot_password',
       name: 'forgot_password',
-      component: Forgot_password,
+      component: ForgotPasswordPage,
     },
   ],
 })
